+++ cloudfit/__init__.py
@@ -15,7 +15,7 @@ from .filter import hard_floor_check, apply_floors
 from .disk import compute_disk_tb, compute_disk_breakdown, list_sequencers
 from .yaml_loader import from_yaml, from_dict
 
-__version__ = "0.2.0"
+__version__ = "0.3.0"
 __author__ = "Chaitanya Krishna Kasaraneni"
 __license__ = "Apache-2.0"